LayoutTests/fast/forms/ValidityState-typeMismatch-datetimelocal-expected.txt

This test aims to check for typeMismatch flag with type=datetime-local input fields

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS "" is a correct valid datetime-local string.
PASS "2009-09-07T16:49" is a correct valid datetime-local string.
PASS "2009-09-07T16:49:31" is a correct valid datetime-local string.
PASS "2009-09-07T16:49:31.1" is a correct valid datetime-local string.
PASS "2009-09-07T16:49:31.12" is a correct valid datetime-local string.
PASS "2009-09-07T16:49:31.123" is a correct valid datetime-local string.
PASS "2009-09-07T16:49:31.1234567890" is a correct valid datetime-local string.
PASS "2147483647-12-31T23:59:59.999" is a correct valid datetime-local string.
PASS "1582-10-15T00:00:00.000" is a correct valid datetime-local string.
PASS " 2009-09-07T16:49 " is a correct invalid datetime-local string.
PASS "2009-09-07t16:49" is a correct invalid datetime-local string.
PASS "2009-09-07 16:49" is a correct invalid datetime-local string.
PASS "2009/09/07T16:49" is a correct invalid datetime-local string.
PASS "a" is a correct invalid datetime-local string.
PASS "-1-09-07T16:49" is a correct invalid datetime-local string.
PASS "1582-10-14T23:59:59" is a correct invalid datetime-local string.
PASS "2147483648-01-01T00:00:00" is a correct invalid datetime-local string.
PASS successfullyParsed is true

TEST COMPLETE
